requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.25
python-multipart==0.0.9
python-dotenv==1.0.1
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
email-validator==2.3.0
